diffusion: removed NonP2P diffusion
