+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/Diffusion/Node/MiniProtocols.hs
@@ -15,6 +15,8 @@ module Test.Ouroboros.Network.Diffusion.Node.MiniProtocols
   , LimitsAndTimeouts (..)
   , AppArgs (..)
   , applications
+    -- * configuration constants
+  , config_REPROMOTE_DELAY
   ) where
 
 import Control.Applicative (Alternative)
@@ -72,8 +74,9 @@ import Ouroboros.Network.Block (HasHeader, HeaderHash, Point)
 import Ouroboros.Network.Block qualified as Block
 import Ouroboros.Network.Context
 import Ouroboros.Network.ControlMessage (ControlMessage (..))
-import Ouroboros.Network.Diffusion.Common qualified as Common
+import Ouroboros.Network.Diffusion.Types qualified as Diff
 import Ouroboros.Network.Driver.Limits
+import Ouroboros.Network.ExitPolicy (RepromoteDelay (..))
 import Ouroboros.Network.KeepAlive
 import Ouroboros.Network.Mock.Chain qualified as Chain
 import Ouroboros.Network.Mock.ConcreteBlock
@@ -84,6 +87,7 @@ import Ouroboros.Network.NodeToNode (blockFetchMiniProtocolNum,
            peerSharingMiniProtocolNum)
 import Ouroboros.Network.NodeToNode.Version (DiffusionMode (..))
 import Ouroboros.Network.PeerSelection.LedgerPeers
+import Ouroboros.Network.PeerSelection.PeerMetric (PeerMetrics)
 import Ouroboros.Network.PeerSelection.PeerSharing qualified as PSTypes
 import Ouroboros.Network.PeerSharing (PeerSharingAPI, bracketPeerSharingClient,
            peerSharingClient, peerSharingServer)
@@ -91,6 +95,7 @@ import Ouroboros.Network.Protocol.PeerSharing.Client (peerSharingClientPeer)
 import Ouroboros.Network.Protocol.PeerSharing.Codec (codecPeerSharing)
 import Ouroboros.Network.Protocol.PeerSharing.Server (peerSharingServerPeer)
 import Ouroboros.Network.Protocol.PeerSharing.Type (PeerSharing)
+import Ouroboros.Network.RethrowPolicy
 import Ouroboros.Network.Util.ShowProxy
 
 import Test.Ouroboros.Network.Diffusion.Node.Kernel
@@ -203,6 +208,8 @@ data AppArgs extraAPI header block m = AppArgs
   , aaChainSyncEarlyExit  :: Bool
   , aaOwnPeerSharing
      :: PSTypes.PeerSharing
+  , aaPeerMetrics
+     :: PeerMetrics m NtNAddr
   }
 
 
@@ -233,9 +240,9 @@ applications :: forall extraAPI block header s m.
              -> LimitsAndTimeouts header block
              -> AppArgs extraAPI header block m
              -> (block -> header)
-             -> Common.Applications NtNAddr NtNVersion NtNVersionData
-                                    NtCAddr NtCVersion NtCVersionData
-                                    extraAPI m ()
+             -> Diff.Applications NtNAddr NtNVersion NtNVersionData
+                                  NtCAddr NtCVersion NtCVersionData
+                                  extraAPI m ()
 applications debugTracer nodeKernel
              Codecs { chainSyncCodec, blockFetchCodec
                     , keepAliveCodec, pingPongCodec
@@ -251,23 +258,39 @@ applications debugTracer nodeKernel
                , aaShouldChainSyncExit
                , aaChainSyncEarlyExit
                , aaOwnPeerSharing
+               , aaPeerMetrics
                }
              toHeader =
-    Common.Applications
-      { Common.daApplicationInitiatorMode =
+    Diff.Applications
+      { Diff.daApplicationInitiatorMode =
           simpleSingletonVersions UnversionedProtocol
                                   (NtNVersionData InitiatorOnlyDiffusionMode aaOwnPeerSharing)
                                   (\NtNVersionData {ntnPeerSharing} -> initiatorApp ntnPeerSharing)
-      , Common.daApplicationInitiatorResponderMode =
+      , Diff.daApplicationInitiatorResponderMode =
           simpleSingletonVersions UnversionedProtocol
                                   (NtNVersionData aaDiffusionMode aaOwnPeerSharing)
                                   (\NtNVersionData {ntnPeerSharing} -> initiatorAndResponderApp ntnPeerSharing)
-      , Common.daLocalResponderApplication =
+      , Diff.daLocalResponderApplication =
           simpleSingletonVersions UnversionedProtocol
                                   UnversionedProtocolData
                                   (\_ -> localResponderApp)
-      , Common.daLedgerPeersCtx =
+      , Diff.daLedgerPeersCtx =
           aaLedgerPeersConsensusInterface
+
+      , Diff.daRethrowPolicy          =
+             muxErrorRethrowPolicy
+          <> ioErrorRethrowPolicy
+
+        -- we are not using local connections, so we can make all the
+        -- errors fatal.
+      , Diff.daLocalRethrowPolicy     =
+             mkRethrowPolicy
+               (\ _ (_ :: SomeException) -> ShutdownNode)
+      , Diff.daPeerMetrics            = aaPeerMetrics
+        -- fetch mode is not used (no block-fetch mini-protocol)
+      , Diff.daBlockFetchMode         = pure (PraosFetchMode FetchModeDeadline)
+      , Diff.daReturnPolicy           = \_ -> config_REPROMOTE_DELAY
+      , Diff.daPeerSharingRegistry    = nkPeerSharingRegistry nodeKernel
       }
   where
     initiatorApp
@@ -608,3 +631,10 @@ applications debugTracer nodeKernel
 
 instance ShowProxy PingPong where
     showProxy Proxy = "PingPong"
+
+--
+-- Constants
+--
+
+config_REPROMOTE_DELAY :: RepromoteDelay
+config_REPROMOTE_DELAY = 10
