--- ouroboros-network/src/Ouroboros/Network/Diffusion/Common.hs
@@ -1,196 +1,0 @@
--- Common things between P2P and NonP2P Diffusion modules
-{-# LANGUAGE DataKinds          #-}
-{-# LANGUAGE GADTs              #-}
-{-# LANGUAGE KindSignatures     #-}
-{-# LANGUAGE RankNTypes         #-}
-{-# LANGUAGE StandaloneDeriving #-}
-
-module Ouroboros.Network.Diffusion.Common
-  ( DiffusionTracer (..)
-  , Failure (..)
-  , Tracers (..)
-  , nullTracers
-  , Arguments (..)
-  , Applications (..)
-  , AcceptedConnectionsLimit (..)
-  ) where
-
-import Control.Concurrent.Class.MonadSTM.Strict
-import Control.Monad.Class.MonadThrow
-import Control.Tracer (Tracer, nullTracer)
-import Data.ByteString.Lazy (ByteString)
-import Data.List.NonEmpty (NonEmpty)
-import Data.Typeable (Typeable)
-import Data.Void (Void)
-
-import Network.Mux qualified as Mx
-
-import Ouroboros.Network.ConnectionId (ConnectionId)
-import Ouroboros.Network.Mux hiding (MiniProtocol (..))
-import Ouroboros.Network.NodeToClient qualified as NodeToClient
-import Ouroboros.Network.NodeToNode (AcceptedConnectionsLimit, DiffusionMode)
-import Ouroboros.Network.NodeToNode qualified as NodeToNode
-import Ouroboros.Network.PeerSelection.Governor.Types (PublicPeerSelectionState)
-import Ouroboros.Network.PeerSelection.LedgerPeers
-           (LedgerPeersConsensusInterface)
-import Ouroboros.Network.Protocol.Handshake
-import Ouroboros.Network.Snocket (FileDescriptor)
-import Ouroboros.Network.Socket (SystemdSocketTracer)
-
--- | The 'DiffusionTracer' logs
---
--- * diffusion initialisation messages
--- * terminal errors thrown by diffusion
---
-data DiffusionTracer ntnAddr ntcAddr
-  = RunServer (NonEmpty ntnAddr)
-  | RunLocalServer ntcAddr
-  | UsingSystemdSocket ntcAddr
-  -- Rename as 'CreateLocalSocket'
-  | CreateSystemdSocketForSnocketPath ntcAddr
-  | CreatedLocalSocket ntcAddr
-  | ConfiguringLocalSocket ntcAddr FileDescriptor
-  | ListeningLocalSocket ntcAddr FileDescriptor
-  | LocalSocketUp  ntcAddr FileDescriptor
-  -- Rename as 'CreateServerSocket'
-  | CreatingServerSocket ntnAddr
-  | ConfiguringServerSocket ntnAddr
-  | ListeningServerSocket ntnAddr
-  | ServerSocketUp ntnAddr
-  -- Rename as 'UnsupportedLocalSocketType'
-  | UnsupportedLocalSystemdSocket ntnAddr
-  -- Remove (this is impossible case), there's no systemd on Windows
-  | UnsupportedReadySocketCase
-  | DiffusionErrored SomeException
-  | SystemdSocketConfiguration SystemdSocketTracer
-    deriving Show
-
--- TODO: add a tracer for these misconfiguration
-data Failure where
-  UnsupportedReadySocket :: Failure
-  UnexpectedIPv4Address  :: forall ntnAddr. (Show ntnAddr, Typeable ntnAddr) => ntnAddr -> Failure
-  UnexpectedIPv6Address  :: forall ntnAddr. (Show ntnAddr, Typeable ntnAddr) => ntnAddr -> Failure
-  NoSocket               :: Failure
-  DiffusionError         :: SomeException -> Failure
-
-deriving instance Show Failure
-instance Exception Failure
-
--- | Common DiffusionTracers interface between P2P and NonP2P
---
-data Tracers ntnAddr ntnVersion ntcAddr ntcVersion m = Tracers {
-      -- | Mux tracer
-      dtMuxTracer
-        :: Tracer m (Mx.WithBearer (ConnectionId ntnAddr) Mx.Trace)
-
-      -- | Handshake protocol tracer
-    , dtHandshakeTracer
-        :: Tracer m (NodeToNode.HandshakeTr ntnAddr ntnVersion)
-
-      --
-      -- NodeToClient tracers
-      --
-
-      -- | Mux tracer for local clients
-    , dtLocalMuxTracer
-        :: Tracer m (Mx.WithBearer (ConnectionId ntcAddr) Mx.Trace)
-
-      -- | Handshake protocol tracer for local clients
-    , dtLocalHandshakeTracer
-        :: Tracer m (NodeToClient.HandshakeTr ntcAddr ntcVersion)
-
-      -- | Diffusion initialisation tracer
-    , dtDiffusionTracer
-        :: Tracer m (DiffusionTracer ntnAddr ntcAddr)
-    }
-
-
-nullTracers :: Applicative m
-            => Tracers ntnAddr ntnVersion
-                       ntcAddr ntcVersion
-                       m
-nullTracers = Tracers {
-    dtMuxTracer            = nullTracer
-  , dtHandshakeTracer      = nullTracer
-  , dtLocalMuxTracer       = nullTracer
-  , dtLocalHandshakeTracer = nullTracer
-  , dtDiffusionTracer      = nullTracer
-  }
-
--- | Common DiffusionArguments interface between P2P and NonP2P
---
-data Arguments m ntnFd ntnAddr ntcFd ntcAddr = Arguments {
-      -- | an @IPv4@ socket ready to accept connections or an @IPv4@ addresses
-      --
-      daIPv4Address              :: Maybe (Either ntnFd ntnAddr)
-
-      -- | an @IPv6@ socket ready to accept connections or an @IPv6@ addresses
-      --
-    , daIPv6Address              :: Maybe (Either ntnFd ntnAddr)
-
-      -- | an @AF_UNIX@ socket ready to accept connections or an @AF_UNIX@
-      -- socket path
-    , daLocalAddress             :: Maybe (Either ntcFd ntcAddr)
-
-      -- | parameters for limiting number of accepted connections
-      --
-    , daAcceptedConnectionsLimit :: AcceptedConnectionsLimit
-
-      -- | run in initiator only mode
-      --
-    , daMode                     :: DiffusionMode
-
-      -- | public peer selection state
-      --
-      -- It is created outside of diffusion, since it is needed to create some
-      -- apps (e.g. peer sharing).
-      --
-    , daPublicPeerSelectionVar   :: StrictTVar m (PublicPeerSelectionState ntnAddr)
-  }
-
-
--- | Versioned mini-protocol bundles run on a negotiated connection.
---
-data Applications ntnAddr ntnVersion ntnVersionData
-                  ntcAddr ntcVersion ntcVersionData
-                  extraAPI m a =
-  Applications {
-      -- | NodeToNode initiator applications for initiator only mode.
-      --
-      -- TODO: we should accept one or the other, but not both:
-      -- 'daApplicationInitiatorMode', 'daApplicationInitiatorResponderMode'.
-      --
-      -- Even in non-p2p mode we use p2p apps.
-      daApplicationInitiatorMode
-        :: Versions ntnVersion
-                    ntnVersionData
-                      (OuroborosBundleWithExpandedCtx
-                      Mx.InitiatorMode ntnAddr
-                      ByteString m a Void)
-
-      -- | NodeToNode initiator & responder applications for bidirectional mode.
-      --
-    , daApplicationInitiatorResponderMode
-           -- Peer Sharing result computation callback
-        :: Versions ntnVersion
-                    ntnVersionData
-                    (OuroborosBundleWithExpandedCtx
-                      Mx.InitiatorResponderMode ntnAddr
-                      ByteString m a ())
-
-      -- | NodeToClient responder application (server role)
-      --
-      -- Because p2p mode does not infect local connections we we use non-p2p
-      -- apps.
-    , daLocalResponderApplication
-        :: Versions ntcVersion
-                    ntcVersionData
-                     (OuroborosApplicationWithMinimalCtx
-                      Mx.ResponderMode ntcAddr
-                      ByteString m Void ())
-
-      -- | Interface used to get peers from the current ledger.
-      --
-      -- TODO: it should be in 'InterfaceExtra'
-    , daLedgerPeersCtx :: LedgerPeersConsensusInterface extraAPI m
-  }
